Move MAX_PEPTIDE_FACETS_OR_DOWNLOAD into config file.
