+++ src/tools/peptide-search/components/results/PeptideSearchResult.tsx
@@ -36,6 +36,7 @@ import {
   LocationToPath,
 } from '../../../../app/config/urls';
 import peptideSearchConverter from '../../adapters/peptideSearchConverter';
+import { MAX_PEPTIDE_FACETS_OR_DOWNLOAD } from '../../config';
 
 import { UniProtkbAPIModel } from '../../../../uniprotkb/adapters/uniProtkbConverter';
 import { SearchResults } from '../../../../shared/types/results';
@@ -93,8 +94,6 @@ enum ServerJobParameters {
   lEqi = 'lEqi',
   swissProtOnly = 'swissProtOnly',
 }
-
-export const MAX_PEPTIDE_FACETS_OR_DOWNLOAD = 1_000;
 
 const PeptideSearchResult = () => {
   const match = useMatchWithRedirect<Params>(
