+++ src/tools/peptide-search/config/index.ts
@@ -1,0 +1,1 @@
+export const MAX_PEPTIDE_FACETS_OR_DOWNLOAD = 1_000;
